feat: requisito 21, cadastra vendas nas tabelas products e salesproducts
