+++ back-end/package.json
@@ -15,8 +15,7 @@
     "test": "mocha src/tests/**/*$NAME*.test.js --exit",
     "test:coverage": "nyc --all npm run test",
     "test:coverage:json": "nyc --all --reporter=json-summary npm run test",
-    "lint": "eslint --no-inline-config --no-error-on-unmatched-pattern -c .eslintrc.json .",
-    "restartBD": "sequelize db:drop && sequelize db:create && sequelize db:migrate && sequelize db:seed:all"
+    "lint": "eslint --no-inline-config --no-error-on-unmatched-pattern -c .eslintrc.json ."
   },
   "dependencies": {
     "@hapi/boom": "9.1.2",
