+++ back-end/src/controller/productSale.js
@@ -1,0 +1,9 @@
+// const { productSale } = require('../service');
+
+// const createSaleProduct = async (req, res) => {
+//   const { productId, saleId, quantity } = req.body;
+//   const register = await productSale.createSalesProducts(productId, saleId, quantity);
+//   return res.status(201).json(register);
+// };
+
+// module.exports = { createSaleProduct };
