+++ back-end/src/controller/saleController.js
@@ -1,14 +1,14 @@
-const { sale } = require('../service');
+const { saleService } = require('../service');
 
 const createSale = async (req, res) => {
   const dataSales = req.body;
-  const Sale = await sale.createSale(dataSales);
-  console.log(sale);
+  const Sale = await saleService.createSale(dataSales);
+  // console.log(sale);
   res.status(201).json(Sale);
 };
 
 const getSales = async (_req, res) => {
-  const sales = await sale.getSales();
+  const sales = await saleService.getSales();
   res.status(200).json(sales);
 };
