+++ back-end/src/routes/index.js
@@ -2,10 +2,12 @@ const login = require('./loginRouter');
 const user = require('./userRouter');
 const product = require('./productRouter');
 const sale = require('./saleRouter');
+// const productSale = require('./productSale');
 
 module.exports = {
   login,
   user,
   product,
   sale,
+  // productSale,
 };
