+++ back-end/src/routes/productSale.js
@@ -1,0 +1,10 @@
+// const express = require('express');
+
+// const { productSales } = require('../controller');
+// const midleware = require('../middleware');
+
+// const router = express.Router();
+
+// router.post('/', productSales.createSaleProduct);
+
+// module.exports = router;
