+++ back-end/src/routes/saleRouter.js
@@ -1,11 +1,11 @@
 const express = require('express');
 
-const controller = require('../controller');
+const { saleController } = require('../controller');
 const midleware = require('../middleware');
 
 const router = express.Router();
 
-router.post('/', midleware.token, controller.sale.createSale);
-router.get('/', controller.sale.getSales);
+router.post('/', midleware.token, saleController.createSale);
+router.get('/', saleController.getSales);
 
 module.exports = router;
