+++ back-end/src/service/productSale.js
@@ -1,0 +1,8 @@
+// const { SalesProduct } = require('../database/models');
+
+// const createSalesProducts = async (productId, saleId, quantity) => {
+//   const register = await SalesProduct.create({ productId, saleId, quantity });
+//   return register;
+// };
+
+// module.exports = { createSalesProducts };
